\name{chk.hy}
\alias{chk.hy}
\title{check whether an object is a hyperSpec object...}
\usage{chk.hy(x)
}
\description{check whether an object is a hyperSpec object}
\details{to be used like validObject}
\alias{chk.hy}
\author{C. Beleites}
\seealso{\code{\link[methods]{validObject}}}
\keyword{methods}
\arguments{\item{x}{the object to check}
}
\examples{chk.hy (chondro)}

