% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_lines.R
\name{count_lines}
\alias{count_lines}
\title{count lines (of an ASCII file)}
\usage{
count_lines(file, chunksize = 10000)
}
\arguments{
\item{file}{the file name or connection}

\item{chunksize}{\code{file} is read in chunks of \code{chunksize} lines.}
}
\value{
number of lines in file
}
\description{
count lines (of an ASCII file)
}
\author{
C. Beleites
}
