% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{.cluster.wavelengths}
\alias{.cluster.wavelengths}
\title{Find clusters of approximately equal wavelengths}
\usage{
.cluster.wavelengths(dots, wl.tolerance)
}
\arguments{
\item{dots}{list of hyperSpec objects to collapse}

\item{wl.tolerance}{wavelength difference tolerance}
}
\value{
data.frame with information about suitable wavelength bins
}
\description{
Find clusters of approximately equal wavelengths
}
