% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{.collapse.equal}
\alias{.collapse.equal}
\title{Try finding groups of hyperSpec objects with (approximately) equal wavelength axes}
\usage{
.collapse.equal(dots, wl.tolerance)
}
\arguments{
\item{dots}{list with hyperSpec object to collapse}

\item{wl.tolerance}{wavelength difference tolerance}
}
\value{
possible shorter list of dots
}
\description{
... and directly rbind.fill them.
}
