\name{dot-DollarNames-dot-hyperSpec}
\alias{.DollarNames.hyperSpec}
\title{command line completion for $}
\usage{.DollarNames.hyperSpec(x, pattern)
}
\description{command line completion for $}
\alias{.DollarNames.hyperSpec}
\author{C. Beleites}
\seealso{\code{\link[utils]{.DollarNames}}}
\keyword{utilities}
\value{the name of the extra data slot}
\arguments{\item{x}{the hyperSpecobject}
\item{pattern}{pattern to look for}
}

