% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_spc_colnames.R
\name{.fix_spc_colnames}
\alias{.fix_spc_colnames}
\title{Ensure that the spectra matrix has the wavelengths in column names}
\usage{
.fix_spc_colnames(spc)
}
\arguments{
\item{spc}{hyperSpec object}
}
\value{
hyperSpec object with wavelengths in column names of \verb{$spc}
}
\description{
Ensure that the spectra matrix has the wavelengths in column names
}
