% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spe.R
\name{.read.spe.xml_string}
\alias{.read.spe.xml_string}
\title{.read.spe.xml_string}
\usage{
.read.spe.xml_string(filename)
}
\arguments{
\item{filename}{- SPE filename}
}
\value{
string containing XML footer
}
\description{
Read XML footer from SPE file format version 3.0 and return it as a long string
for subsequent parsing. Basically the purpose of this function is to check
that the file format version is 3.0 or above, and to find and read the
correct part of this file.
}
