% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplevels.R
\docType{methods}
\name{droplevels,hyperSpec-method}
\alias{droplevels,hyperSpec-method}
\title{droplevels for hyperSpec object}
\usage{
\S4method{droplevels}{hyperSpec}(x, ...)
}
\arguments{
\item{x}{hyperSpec object}

\item{...}{handed to \link[base]{droplevels}}
}
\value{
hyperSpec object with unused levels of all factors in @data dropped.
}
\description{
calls droplevels on the data.frame in spc@data.
}
\details{
Use parameter \code{except} to exclude columns from dropping.
}
\examples{

chondro[1:3]$clusters
droplevels (chondro [1:3])$clusters
}
\seealso{
\link[base]{droplevels}
}
