% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fauxCell.R
\docType{data}
\name{fauxCell}
\alias{fauxCell}
\title{Faux Cell Data Set for Testing & Demonstration}
\format{
The object contains 875 Raman-like spectra, allocated to three
  groups/regions in column region: the matrix/background, the cell and the
  cell nucleus. Each spectrum is composed of 300 data points.  The spectrum
  of each region is unique and simple, with a single peak at a particular
  frequency and line width.  Poisson noise has been added.  The data is
  indexed along the x and y dimensions, simulating data collected on a grid.
}
\usage{
fauxCell
}
\description{
This is a synthetic data set intended for testing and demonstration.
}
\details{
The data set resembles the `chondro` data set but is entirely synthetic.
}
\examples{

fauxCell

plot (sample (fauxCell, 10), stacked = TRUE)

# Plot mean spectra
FCgrps <- aggregate(fauxCell, fauxCell$region, mean_pm_sd)
plotspc(FCgrps, stacked = ".aggregate",
        col = c("red", "green", "blue"), fill = ".aggregate")

mapcols <- c(cell = "aquamarine", matrix = "aliceblue", nucleus = "dodgerblue")
plotmap(fauxCell, region ~ x * y, col.regions = mapcols)

# PCA
pca <- prcomp(fauxCell)
plot(pca)

loadings <- decomposition(fauxCell, t(pca$rotation), scores = FALSE)
plot(loadings[1 : 5], stacked = TRUE)

plot(pca$x[,2], pca$x[,3], xlab = "PC 1", ylab = "PC 2",
  bg = mapcols[fauxCell$region], col = "black", pch = 21)

}
\author{
Claudia Beleites, Bryan A. Hanson
}
\keyword{datasets}
