\name{flu}
\alias{flu}
\docType{data}
\title{Chitin Fluorescence Spectra}
\description{
Fluorescence spectra of different dilutions of chitine forming a
calibration set.
}
\usage{data(flu)}
\format{
 The data set has 6 fluorescence emission spectra measured on
 concentrations between 0.05 mg/l and 0.30 mg/l.
 Each spectrum has 181 data points in the range of405nm to 495 nm. 
}
\details{
  see \link{Vignette}
}
\author{M. Kammer and C. Beleites}
\examples{
flu

plot (flu)

plotc (flu)
}
\keyword{datasets}
