% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEW-functions.R
\name{Future-functions}
\alias{Future-functions}
\alias{.spc_io_postprocess_optional}
\alias{wl_convert_units}
\alias{hy_set_options}
\alias{hy_get_option}
\alias{read_txt_long}
\alias{read_txt_wide}
\alias{.wl_fix_unit_name}
\alias{assert_hyperSpec}
\title{Future functions}
\usage{
.spc_io_postprocess_optional(...)

wl_convert_units(x, from, to, ref_wl = NULL)

hy_set_options(...)

hy_get_option(...)

hy_set_options(...)

read_txt_long(...)

read_txt_wide(...)

.wl_fix_unit_name(...)

assert_hyperSpec(...)
}
\arguments{
\item{...}{Arguments to functions.}

\item{x, from, to, ref_wl}{Arguments to functions.}
}
\description{
These functions will be introduced in \pkg{hyperSpec} v1.0 and will replace
some current functions. Now they appear here just for compatibility with
other packages, which should be released on CRAN. They are not intended to
be used by \pkg{hyperSpec} v0.100 users directly.
}
