\name{math}
\alias{log,-method}
\alias{log}
\title{log,-method}
\alias{log}
\alias{log,hyperSpec-method}
\alias{Math2,-method}
\alias{Math2}
\description{Math Functions for hyperSpec Objects}
\details{\code{Math2,-method}: The functions \code{abs}, \code{sign},

\code{sqrt},

\code{floor}, \code{ceiling}, \code{trunc}, \code{round}, \code{signif},

\code{exp}, \code{log}, \code{expm1}, \code{log1p},

\code{cos}, \code{sin}, \code{tan}, \code{acos}, \code{asin}, \code{atan},
\code{cosh}, \code{sinh}, \code{tanh}, \code{acosh}, \code{asinh},
\code{atanh},

\code{lgamma}, \code{gamma}, \code{digamma}, \code{trigamma},

\code{cumsum}, \code{cumprod}, \code{cummax}, \code{cummin}

for \code{hyperSpec} objects.

}
\alias{Math}
\alias{Math2}
\alias{Math,hyperSpec-method}
\alias{Math2,hyperSpec-method}
\alias{abs,hyperSpec-method}
\alias{sign,hyperSpec-method}
\alias{sqrt,hyperSpec-method}
\alias{floor,hyperSpec-method}
\alias{ceiling,hyperSpec-method}
\alias{trunc,hyperSpec-method}
\alias{round,hyperSpec-method}
\alias{signif,hyperSpec-method}
\alias{exp,hyperSpec-method}
\alias{log,hyperSpec-method}
\alias{expm1,hyperSpec-method}
\alias{log1p,hyperSpec-method}
\alias{cos,hyperSpec-method}
\alias{sin,hyperSpec-method}
\alias{tan,hyperSpec-method}
\alias{acos,hyperSpec-method}
\alias{asin,hyperSpec-method}
\alias{atan,hyperSpec-method}
\alias{cosh,hyperSpec-method}
\alias{sinh,hyperSpec-method}
\alias{tanh,hyperSpec-method}
\alias{acosh,hyperSpec-method}
\alias{asinh,hyperSpec-method}
\alias{atanh,hyperSpec-method}
\alias{lgamma,hyperSpec-method}
\alias{gamma,hyperSpec-method}
\alias{digamma,hyperSpec-method}
\alias{trigamma,hyperSpec-method}
\alias{cumsum,hyperSpec-method}
\alias{cumprod,hyperSpec-method}
\alias{cummax,hyperSpec-method}
\alias{cummin,hyperSpec-method}
\alias{round,hyperSpec-method}
\alias{signif,hyperSpec-method}
\value{\code{Math2,-method}: a \code{hyperSpec} object

}
\author{C. Beleites}
\seealso{\code{\link[methods]{S4groupGeneric}} for group generic methods.

\code{\link[base]{Math}} for the base math functions.

\code{\link[hyperSpec]{Arith}} for arithmetic operators,
\code{\link[hyperSpec]{Comparison}} for comparison operators, and
\code{\link[hyperSpec]{Summary}} for group generic functions working on
\code{hyperSpec} objects.}
\keyword{methods}
\keyword{math}
\arguments{\item{x}{the \code{hyperSpec} object}
\item{digits}{integer stating the rounding precision}
}
\examples{
log (flu)
}
\alias{log,-method}
\alias{log}
\alias{log}
\alias{log,hyperSpec-method}
\alias{Math,-method}
\alias{Math}

