% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeextra.R
\name{mergeextra}
\alias{mergeextra}
\title{Merge additional extra data into an object}
\usage{
mergeextra(x, y)
}
\arguments{
\item{x}{hyperSpec object into which extra data columns of y are to be merged}

\item{y}{hyperSpec object with further extra data columns}
}
\value{
x amended with columns from y
}
\description{
Columns from `y` that are not in `x` are merged into `x`. Columns in `x` and
`y` are considered equal if they have the same name, the same content, and
the same label.
}
\details{
Columns with different content than a column in `x` with the same name will
get `".y"` appended to its name.

**`mergeextra()` relies on `x` and `y` having the same row order.**
}
\examples{
tmp <- flu [,FALSE,]
tmp$prediction <- 1:6
tmp

mergeextra(flu, tmp)
}
