\name{paracetamol}
\alias{paracetamol}
\alias{paracetamol}
\title{Paracetamol Spectrum...}
\description{Paracetamol Spectrum
A Raman spectrum of a paracetamol tablet.}
\docType{data}
\author{C. Beleites}
\keyword{datasets}
\examples{
paracetamol

plot (paracetamol)
plotspc (paracetamol, c (min ~ 1750, 2800 ~ max), xoffset = 800,
wl.reverse = TRUE)
}

