% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/read.asc.Andor.R
\name{scan.asc.Andor}
\alias{scan.asc.Andor}
\alias{read.asc.Andor}
\title{File Import Andor Solis}
\usage{
scan.asc.Andor(...)

read.asc.Andor(
  file = stop("filename or connection needed"),
  ...,
  quiet = TRUE,
  dec = ".",
  sep = ","
)
}
\arguments{
\item{..., quiet, dec, sep}{handed to \code{\link[base]{scan}}}

\item{file}{filename or connection to ASCII file}
}
\value{
a hyperSpec object
}
\description{
Import Raman Spectra/Maps from Andor Cameras/Solis ASCII files
}
\details{
\code{read.asc.Andor} reads Andor Solis ASCII (\code{.asc}) files where the first column gives the wavelength
axes and the other columns the spectra.
}
\seealso{
\code{vignette ("fileio")} for more information on file import and

\code{\link{options}} for details on options.
}
\author{
Claudia Beleites
}
\keyword{internal}
