% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.asc.PerkinElmer.R
\name{read.asc.PerkinElmer}
\alias{read.asc.PerkinElmer}
\title{File import filter PerkinElmer ASCII spectra}
\usage{
read.asc.PerkinElmer(file = stop("filename or connection needed"), ...)
}
\arguments{
\item{file}{filename (or connection)}

\item{...}{further parameters are handed to \code{\link[hyperSpec]{read.txt.long}}}
}
\value{
hyperSpec object
}
\description{
Imports a single spectrum in PerkinElmer's ASCII format. This function is experimental.
}
