% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unittest.R
\name{hy.unittest}
\alias{hy.unittest}
\title{hyperSpec unit tests}
\usage{
hy.unittest(standalone = TRUE, reporter = "progress")
}
\arguments{
\item{standalone}{run the unit test on their own, e.g. from the console
(`TRUE`) or within testthat tests (`FALSE`), e.g. via `devtools::test()`}

\item{reporter}{the reporter to use, defaults to [testthat::ProgressReporter]}
}
\value{
Invisibly returns a data frame with the test results
}
\description{
If \code{\link[testthat]{testthat}} is available, run the unit tests and
display the results.
}
\examples{

hy.unittest ()

}
\author{
Claudia Beleites
}
\keyword{programming}
\keyword{utilities}
