% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wc}
\alias{wc}
\title{wc
word count of ASCII files}
\usage{
wc(file, flags = c("lines", "words", "bytes"))
}
\arguments{
\item{file}{the file name or pattern}

\item{flags}{the parameters to count, character vector with the long form
of the parameters}
}
\value{
data.frame with the counts and file names, or \code{NULL} if wc is
  not available
}
\description{
wc uses the system command wc
}
\author{
C. Beleites
}

