% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.hy.R
\name{chk.hy}
\alias{chk.hy}
\alias{validObject}
\alias{validObject,hyperSpec-method}
\title{Validation of hyperSpec objects}
\usage{
chk.hy(object)
}
\arguments{
\item{object}{the object to check}
}
\value{
\code{TRUE} if the check passes, otherwise stop with an
error.
}
\description{
Check whether an object is a hyperSpec object and validate the object
}
\examples{
chk.hy (chondro)
validObject (chondro)
}
\seealso{
\code{\link[methods]{validObject}}
}
\author{
C. Beleites
}
\keyword{methods}
