% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMeans.R
\name{colSums}
\alias{colSums}
\alias{colMeans,hyperSpec-method}
\alias{colSums,hyperSpec-method}
\alias{rowMeans,hyperSpec-method}
\alias{rowSums,hyperSpec-method}
\title{colSums, colMeans, rowSums and rowMeans functions for hyperSpec objects}
\usage{
\S4method{colMeans}{hyperSpec}(x, na.rm = TRUE, ..., label.spc)

\S4method{colSums}{hyperSpec}(x, na.rm = TRUE, ..., label.spc)

\S4method{rowMeans}{hyperSpec}(x, na.rm = TRUE, ..., label.wavelength)

\S4method{rowSums}{hyperSpec}(x, na.rm = TRUE, ..., label.wavelength)
}
\arguments{
\item{x}{hyperSpec object}

\item{na.rm, ...}{further parameters to the base functions

\code{na.rm} defaults to \code{TRUE} for hyperSpec objects.}

\item{label.spc}{labels for the intensity axis for loadings-like (col) statistics}

\item{label.wavelength}{labels for the wavelength axis for scores-like (row) statistics}
}
\description{
hyperSpec objects can use the base functions \code{\link[base]{colMeans}},
\code{\link[base]{colSums}}, \code{\link[base]{rowMeans}} and \code{\link[base]{rowSums}}.
}
\examples{
colMeans (flu)
colSums (flu)
colSums (flu)
rowSums (flu)
}
\seealso{
\link[base]{colSums}
}
