% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\docType{methods}
\name{ncol,hyperSpec-method}
\alias{ncol,hyperSpec-method}
\alias{nrow,hyperSpec-method}
\alias{nwl}
\alias{dim,hyperSpec-method}
\alias{length,hyperSpec-method}
\title{The Number of Rows (Spectra), Columns, and Data Points per Spectrum of an
hyperSpec Object)}
\usage{
\S4method{ncol}{hyperSpec}(x)

\S4method{nrow}{hyperSpec}(x)

nwl(x)

\S4method{dim}{hyperSpec}(x)

\S4method{length}{hyperSpec}(x)
}
\arguments{
\item{x}{a \code{hyperSpec} object}
}
\value{
\code{nrow}, \code{ncol}, \code{nwl}, and \code{length}, return an
  \code{integer}.

\code{dim} returns a vector of length 3.
}
\description{
\code{ncol} returns the number of columns in \code{x@data}. I.e. the number
of columns with additional information to each spectrum (e.g. "x", "y",
\dots{}) + 1 (for column \code{spc} containing the spectra).
}
\examples{

ncol (chondro)
nrow (chondro)

nwl  (chondro)
dim (chondro)
length (chondro)
}
\seealso{
\code{\link[base]{ncol}}

\code{\link[base]{nrow}}

\code{\link[base]{dim}}

\code{\link[base]{length}}
}
\author{
C. Beleites
}
\keyword{methods}
