% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DollarNames.R
\name{.DollarNames.hyperSpec}
\alias{.DollarNames.hyperSpec}
\alias{.DollarNames}
\alias{.DollarNames,hyperSpec-method}
\title{command line completion for $}
\usage{
\method{.DollarNames}{hyperSpec}(x, pattern = "")
}
\arguments{
\item{x}{the hyperSpecobject}

\item{pattern}{pattern to look for}
}
\value{
the name of the extra data slot
}
\description{
command line completion for $
}
\seealso{
\code{\link[utils]{.DollarNames}}
}
\author{
C. Beleites
}
\keyword{utilities}
