\name{.DollarNames.hyperSpec}
\alias{.DollarNames}
\alias{.DollarNames.hyperSpec}
\alias{.DollarNames,hyperSpec-method}
\title{command line completion for $}
\usage{
  \method{.DollarNames}{hyperSpec} (x, pattern)
}
\arguments{
  \item{x}{the hyperSpecobject}

  \item{pattern}{pattern to look for}
}
\value{
  the name of the extra data slot
}
\description{
  command line completion for $
}
\author{
  C. Beleites
}
\seealso{
  \code{\link[utils]{.DollarNames}}
}
\keyword{utilities}

