% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplevels.R
\name{droplevels,hyperSpec-method}
\alias{droplevels,hyperSpec-method}
\title{droplevels for hyperSpec object}
\usage{
\S4method{droplevels}{hyperSpec}(x, ...)
}
\arguments{
\item{x}{hyperSpec object}

\item{...}{handed to \code{\link[base:droplevels.data.frame]{base::droplevels.data.frame()}}}
}
\value{
hyperSpec object with unused levels of all factors in \verb{@data} dropped.
}
\description{
calls \code{\link[base:droplevels]{base::droplevels()}} on the data.frame in \code{spc@data}.
}
\examples{

chondro[1:3]$clusters
droplevels(chondro [1:3])$clusters
}
\seealso{
\code{\link[base:droplevels]{base::droplevels()}}
}
