% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flu.R
\docType{data}
\name{flu}
\alias{flu}
\title{Quinine Fluorescence Spectra
Fluorescence spectra of different dilutions of quinine forming a
calibration set.}
\format{
The data set has 6 fluorescence emission spectra measured on
  quinine concentrations between 0.05 mg/l and 0.30 mg/l.  Each spectrum
  consists of 181 data points in the range of 405 nm to 495 nm.
}
\description{
See the vignette: \code{vignette ("flu", package = "hyperSpec")}
}
\examples{

flu

plot (flu)

plotc (flu)

}
\author{
M. Kammer and C. Beleites
}
\keyword{datasets}
