\name{flu}
\alias{flu}
\alias{flu}
\title{Quinine Fluorescence Spectra...}
\description{Quinine Fluorescence Spectra
Fluorescence spectra of different dilutions of quinine forming a
calibration set.}
\details{See the vignette.}
\docType{data}
\author{M. Kammer and C. Beleites}
\keyword{datasets}
\examples{
flu

plot (flu)

plotc (flu)

\dontrun{vignette ("flu", package = "hyperSpec")}
}

