\docType{data}
\name{flu}
\alias{flu}
\title{Quinine Fluorescence Spectra
Fluorescence spectra of different dilutions of quinine forming a
calibration set.}
\format{The data set has 6 fluorescence emission spectra measured on
  quinine concentrations between 0.05 mg/l and 0.30 mg/l.  Each spectrum
  consists of 181 data points in the range of 405 nm to 495 nm.}
\description{
  See the vignette.
}
\examples{
flu

plot (flu)

plotc (flu)

\dontrun{vignette ("flu", package = "hyperSpec")}
}
\author{
  M. Kammer and C. Beleites
}
\keyword{datasets}

