% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplotmixmap.R
\name{legendright}
\alias{legendright}
\alias{qmixtile}
\alias{normalize.colrange}
\alias{normalize.range}
\alias{normalize.null}
\alias{normalize.minmax}
\alias{qmixlegend}
\alias{colmix.rgb}
\title{Plot multivariate data into colour channels}
\usage{
legendright(p, l, legend.width = 8, legend.unit = "lines")

qmixtile(
  object,
  purecol = stop("pure component colors needed."),
  mapping = aes_string(x = "x", y = "y", fill = "spc"),
  ...,
  map.tileonly = FALSE
)

normalize.colrange(x, na.rm = TRUE, legend = FALSE, n = 100, ...)

normalize.range(x, na.rm = TRUE, legend = FALSE, n = 100, ...)

normalize.null(x, na.rm = TRUE, legend = FALSE, n = 100, ...)

normalize.minmax(x, min = 0, max = 1, legend = FALSE, n = 100, ...)

qmixlegend(
  x,
  purecol,
  dx = 0.33,
  ny = 100,
  labels = names(purecol),
  normalize = normalize.colrange,
  ...
)

colmix.rgb(
  x,
  purecol,
  against = 1,
  sub = TRUE,
  normalize = normalize.colrange,
  ...
)
}
\arguments{
\item{p}{plot object}

\item{l}{legend object}

\item{legend.width, legend.unit}{size of legend part}

\item{object}{matrix to be plotted with mixed colour channels}

\item{purecol}{pure component colours, names determine legend labels}

\item{mapping}{see [ggplot2::geom_tile()]}

\item{...}{`qmixtile`: handed to [colmix.rgb()]

  `qmixlegend()` and `colmix.rgb()` hand further arguments to the
  `normalize` function}

\item{map.tileonly}{if `TRUE`, `mapping` will be handed to
[ggplot2::geom_tile()] instead of [ggplot2::ggplot()].}

\item{x}{matrix with component intensities in columns}

\item{na.rm}{see [base::min()]}

\item{legend}{should a legend be produced instead of normalized values?}

\item{n}{of colours to produce in legend}

\item{min}{numeric with value corresponding to "lowest" colour for each
column}

\item{max}{numeric with value corresponding to "hightest" colour for each
column}

\item{dx}{width of label bar}

\item{ny}{number of colours in legend}

\item{labels}{component names}

\item{normalize}{function to normalize the values.}

\item{against}{value to mix against
(for `sub = TRUE` only, 1 = white, 0 = black)}

\item{sub}{subtractive color mixing?}
}
\value{
invisible `NULL`

list with components `ymin`, `max` and `fill` to specify value and
  fill colour value (still numeric!) for the legend, otherwise the
  normalized values

ggplot object with legend

character with colours
}
\description{
plot graph with legend right of it
}
\author{
Claudia Beleites

Claudia Beleites

Claudia Beleites
}
