% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplotmixmap.R
\name{qplotmixmap}
\alias{qplotmixmap}
\title{qplotmap with colour mixing for multivariate overlay}
\usage{
qplotmixmap(object, ...)
}
\arguments{
\item{object}{hyperSpec object}

\item{...}{handed over to \code{\link[hyperSpec]{qmixlegend}} and \code{\link[hyperSpec]{qmixtile}}}
}
\value{
invisible list with ggplot2 objects map and legend
}
\description{
map plot with colour overlay.
}
\examples{
chondro <- chondro - spc.fit.poly.below (chondro)
chondro <- sweep (chondro, 1, apply (chondro, 1, mean), "/")
chondro <- sweep (chondro, 2, apply (chondro, 2, quantile, 0.05), "-")

qplotmixmap (chondro [,,c (940, 1002, 1440)],
             purecol = c (colg = "red", Phe = "green", Lipid = "blue"))

}
\seealso{
\code{\link[hyperSpec]{qmixtile}}
}
\author{
Claudia Beleites
}
