% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ini.R
\name{read.ini}
\alias{read.ini}
\title{Read INI files}
\usage{
read.ini(
  con = stop("Connection con needed."),
  skip = NULL,
  encoding = "unknown"
)
}
\arguments{
\item{con}{connection or file name}

\item{skip}{number of lines to skip before first \code{[section]} starts}

\item{encoding}{see \code{\link[base]{readLines}}}
}
\value{
a list with one element per section in the .ini file, each containing a list with elements
for the key-value-pairs.
}
\description{
\code{read.ini} reads ini files of the form
}
\details{
[section]
key = value

into a list.

\code{read.ini} sanitizes the element names and tries to convert scalars and comma separated
numeric vectors to numeric.
}
\author{
C. Beleites
}
\keyword{IO}
\keyword{file}
