% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/read.mat.Cytospec.R
\name{read.cytomat}
\alias{read.cytomat}
\alias{read.mat.Cytospec}
\title{Import for Cytospec mat files}
\usage{
read.cytomat(...)

read.mat.Cytospec(file, keys2data = FALSE, blocks = TRUE)
}
\arguments{
\item{...}{\code{read.cytomat} for now hands all arguments to
\code{read.mat.Cytospec} for backwards compatibility.}

\item{file}{The complete file name (or a connection to) the .mat file.}

\item{keys2data}{specifies which elements of the \code{Info} should be
transferred into the extra data}

\item{blocks}{which blocks should be read? \code{TRUE} reads all blocks.}
}
\value{
hyperSpec object if the file contains a single spectra block,
  otherwise a list with one hyperSpec object for each block.
}
\description{
These functions allow to import .mat (Matlab V5) files written by Cytospec.
}
\details{
\code{read.cytomat} has been renamed to \code{read.mat.Cytospec} and is now
deprecated. Use \code{read.mat.Cytospec} instead.
}
\note{
This function is an ad-hoc implementation and subject to changes.
}
\seealso{
\code{R.matlab::readMat}
}
\author{
C. Beleites
}
\keyword{IO}
\keyword{file}
