% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.txt.Shimadzu.R
\name{read.txt.Shimadzu}
\alias{read.txt.Shimadzu}
\title{Reads Shimadzu GCxGC-qMS - Spectra Files (.txt) as exported by Shimadzu Chrome Solution (v. 2.72)
Mass Spectrometer: Shimadzu GCMS-QP 2010 Ultra (www.shimadzu.com)}
\usage{
read.txt.Shimadzu(filename, encoding = "", quiet = TRUE)
}
\arguments{
\item{filename}{file name and path of the .txt file}

\item{encoding}{encoding of the txt file (used by \code{\link[base]{readLines}})}

\item{quiet}{suppress printing of progress}
}
\value{
list of spectra tables
}
\description{
Reads Shimadzu GCxGC-qMS - Spectra Files (.txt) as exported by Shimadzu Chrome Solution (v. 2.72)
Mass Spectrometer: Shimadzu GCMS-QP 2010 Ultra (www.shimadzu.com)
}
\note{
This is a first rough import function and the functions may change without notice.
}
\author{
Bjoern Egert
}
