% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\docType{methods}
\name{subset}
\alias{subset}
\alias{subset,hyperSpec-method}
\title{subset}
\usage{
\S4method{subset}{hyperSpec}(x, ...)
}
\arguments{
\item{x}{hyperSpec object}

\item{...}{handed to \code{\link[base]{subset}} (data.frame method)}
}
\value{
hyperSpec object containing the respective subset of spectra.
}
\description{
subset for hyperSpec object
}
\seealso{
\code{\link[base]{subset}}
}
\author{
Claudia Beleites
}
