% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\docType{methods}
\name{Summary}
\alias{Summary}
\alias{Summary,hyperSpec-method}
\alias{all,hyperSpec-method}
\alias{any,hyperSpec-method}
\alias{sum,hyperSpec-method}
\alias{prod,hyperSpec-method}
\alias{min,hyperSpec-method}
\alias{max,hyperSpec-method}
\alias{range,hyperSpec-method}
\alias{is.na,hyperSpec-method}
\alias{all_wl}
\alias{any_wl}
\title{The functions}
\usage{
\S4method{Summary}{hyperSpec}(x, ..., na.rm = FALSE)

\S4method{is.na}{hyperSpec}(x)

all_wl(expression, na.rm = FALSE)

any_wl(expression, na.rm = FALSE)
}
\arguments{
\item{x}{hyperSpec object}

\item{...}{further objects}

\item{na.rm}{logical indicating whether missing values should be removed}

\item{expression}{expression that evaluates to a logical matrix of the same size as the spectra matrix}
}
\value{
\code{sum}, \code{prod}, \code{min}, \code{max}, and \code{range} return  a numeric,
\code{all}, \code{any}, and \code{is.na} a logical.
}
\description{
\code{all}, \code{any},
}
\details{
\code{sum}, \code{prod},

\code{min}, \code{max},

\code{range}, and

\code{is.na}

for \code{hyperSpec} objects.

All these functions work on the spectra matrix.
}
\examples{

	range (flu)


is.na (flu [,, 405 ~ 410]);

all_wl (flu > 100)

any_wl (flu > 300)
! any_wl (is.na (flu))
}
\seealso{
\code{\link[base]{Summary}} for the base summary functions.

\code{\link[base]{all.equal}} and \code{\link[base]{isTRUE}}
}
