% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unittest.R
\name{hy.unittest}
\alias{hy.unittest}
\title{hyperSpec unit tests}
\usage{
hy.unittest()
}
\value{
Invisibly returns a data frame with the test results
}
\description{
If \code{\link[testthat]{testthat}} is available, run the unit tests and
display the results.
}
\examples{

hy.unittest ()

}
\author{
Claudia Beleites
}
\keyword{programming}
\keyword{utilities}
