% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vandermonde.R
\name{vanderMonde}
\alias{vanderMonde}
\alias{vanderMonde,hyperSpec-method}
\title{Function evaluation on hyperSpec objects}
\usage{
vanderMonde(x, order, ...)

\S4method{vanderMonde}{hyperSpec}(x, order, ..., normalize.wl = normalize01)
}
\arguments{
\item{x}{object to evaluate the polynomial on}

\item{order}{of the polynomial}

\item{...}{hyperSpec method: further arguments to `hyperSpec::decomposition()`}

\item{normalize.wl}{function to transorm the wavelengths before evaluating the polynomial (or
other function). `hyperSpec::normalize01()` maps the wavelength range to the interval
[0, 1]. Use `base::I()` to turn off.}
}
\value{
Vandermonde matrix

hyperSpec method: hyperSpec object containing van der Monde matrix as spectra and an additional column `$.vdm.order$ giving the order of each spectrum (term).
}
\description{
\code{vandermonde()} generates Vandermonde matrices, the hyperSpec method
generates a hyperSpec object containing the van der Monde matrix of the
wavelengths of a hyperSpec object.
}
\details{
It is often numerically preferrable to map \code{wl(x)} to [0, 1], see the
example.
}
\examples{
plot(vanderMonde(flu, 2))
plot(vanderMonde(flu, 2, normalize.wl = I))

}
\seealso{
`hyperSpec::wl.eval()` for calculating arbitrary functions of the wavelength,

`hyperSpec::normalize01()`
}
\author{
C. Beleites
}
