% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc.R
\name{wc}
\alias{wc}
\title{line/word/character count of ASCII files}
\usage{
wc()
}
\description{
`wc()` is defunct and will be removed from hyperSpec in future. Consider using [count_lines()] instead for line counting.
}
\seealso{
[count_lines()]
}
\author{
C. Beleites
}
